+++ app/Domain/CatchEmAll/Achievements/Utils/AchievementRegister.php
@@ -3,13 +3,14 @@
 namespace App\Domain\CatchEmAll\Achievements\Utils;
 
 use App\Domain\CatchEmAll\Achievements\Archivist;
-use App\Domain\CatchEmAll\Achievements\BugBountyHunter;
 use App\Domain\CatchEmAll\Achievements\Collector;
 use App\Domain\CatchEmAll\Achievements\Curator;
 use App\Domain\CatchEmAll\Achievements\FirstCatch;
 use App\Domain\CatchEmAll\Achievements\FuredexComplete;
 use App\Domain\CatchEmAll\Achievements\GottaCatchEmAll;
 use App\Domain\CatchEmAll\Achievements\Nice;
+use App\Domain\CatchEmAll\Achievements\Special\BugBountyHunter;
+use App\Domain\CatchEmAll\Achievements\Special\CEATeam;
 use App\Domain\CatchEmAll\Achievements\TheLegendary151;
 use App\Domain\CatchEmAll\Enums\SpecialCodeType;
 use App\Domain\CatchEmAll\Interface\Achievement;
@@ -75,24 +76,18 @@ class AchievementRegister
     /**
      * Count of required (non-optional) achievements for 100% completion.
      * Built during initialization.
-     *
-     * @var int
      */
     protected static int $requiredAchievementCount = 0;
 
     /**
      * Count of optional achievements.
      * Built during initialization.
-     *
-     * @var int
      */
     protected static int $optionalAchievementCount = 0;
 
     /**
      * Initialize the achievement register.
      * This method is called once during application startup.
-     *
-     * @return void
      */
     public static function init(): void
     {
@@ -109,7 +104,7 @@ class AchievementRegister
         self::calculateAchievementCounts();
 
         // Log initialization
-        Log::info('AchievementRegister initialized with ' . count(self::$achievements) . ' achievements', [
+        Log::info('AchievementRegister initialized with '.count(self::$achievements).' achievements', [
             'total_achievements' => count(self::$achievements),
             'special_achievements' => count(self::$specialCodeIndex),
             'normal_achievements' => count(self::$normalAchievements),
@@ -120,23 +115,19 @@ class AchievementRegister
 
     /**
      * Build achievement instances from registered classes.
-     *
-     * @return void
      */
     protected static function buildAchievementInstances(): void
     {
         self::$achievements = [];
 
         foreach (self::$achievementClasses as $className) {
-            self::$achievements[$className] = new $className();
+            self::$achievements[$className] = new $className;
         }
     }
 
     /**
      * Build all indexes for fast lookups.
      * Called during initialization.
-     *
-     * @return void
      */
     protected static function buildIndexes(): void
     {
@@ -147,8 +138,6 @@ class AchievementRegister
 
     /**
      * Build the ID => Achievement index.
-     *
-     * @return void
      */
     protected static function buildIdIndex(): void
     {
@@ -162,8 +151,6 @@ class AchievementRegister
 
     /**
      * Build the SpecialCodeType => SpecialAchievement index.
-     *
-     * @return void
      */
     protected static function buildSpecialCodeIndex(): void
     {
@@ -174,7 +161,7 @@ class AchievementRegister
                 $specialCode = $achievement->getSpecialCode();
                 $codeValue = $specialCode->name;
 
-                if (!isset(self::$specialCodeIndex[$codeValue])) {
+                if (! isset(self::$specialCodeIndex[$codeValue])) {
                     self::$specialCodeIndex[$codeValue] = [];
                 }
 
@@ -185,15 +172,13 @@ class AchievementRegister
 
     /**
      * Build the normal achievements index (non-special achievements).
-     *
-     * @return void
      */
     protected static function buildNormalAchievementsIndex(): void
     {
         self::$normalAchievements = [];
 
         foreach (self::$achievements as $achievement) {
-            if (!($achievement instanceof SpecialAchievement)) {
+            if (! ($achievement instanceof SpecialAchievement)) {
                 self::$normalAchievements[] = $achievement;
             }
         }
@@ -202,8 +187,6 @@ class AchievementRegister
     /**
      * Calculate achievement counts for required and optional achievements.
      * Called during initialization.
-     *
-     * @return void
      */
     protected static function calculateAchievementCounts(): void
     {
@@ -221,9 +204,6 @@ class AchievementRegister
 
     /**
      * Get an achievement by its ID using the index for fast lookup.
-     *
-     * @param string $achievementId
-     * @return Achievement|null
      */
     public static function getAchievementById(string $achievementId): ?Achievement
     {
@@ -233,7 +213,6 @@ class AchievementRegister
     /**
      * Get all special achievements that can be triggered by a specific SpecialCodeType.
      *
-     * @param SpecialCodeType $specialCode
      * @return array<SpecialAchievement>
      */
     public static function getAchievementsBySpecialCode(SpecialCodeType $specialCode): array
@@ -264,7 +243,6 @@ class AchievementRegister
     /**
      * Get special achievements by their special code.
      *
-     * @param SpecialCodeType $specialCode
      * @return array<SpecialAchievement>
      */
     public static function getSpecialAchievementsByCode(SpecialCodeType $specialCode): array
@@ -295,8 +273,7 @@ class AchievementRegister
     /**
      * Get a specific achievement instance by its class name.
      *
-     * @param class-string<Achievement> $className
-     * @return Achievement|null
+     * @param  class-string<Achievement>  $className
      */
     public static function getAchievement(string $className): ?Achievement
     {
@@ -305,8 +282,6 @@ class AchievementRegister
 
     /**
      * Get the number of required (non-optional) achievements for 100% completion.
-     *
-     * @return int
      */
     public static function getRequiredAchievementCount(): int
     {
@@ -315,8 +290,6 @@ class AchievementRegister
 
     /**
      * Get the number of optional achievements.
-     *
-     * @return int
      */
     public static function getOptionalAchievementCount(): int
     {
@@ -325,9 +298,6 @@ class AchievementRegister
 
     /**
      * Check if an achievement is registered by ID.
-     *
-     * @param string $achievementId
-     * @return bool
      */
     public static function hasAchievementId(string $achievementId): bool
     {
@@ -353,20 +323,19 @@ class AchievementRegister
     /**
      * Validate all registered achievements for consistency.
      *
-     * @return void
      * @throws \InvalidArgumentException
      */
     private static function validateAchievements(): void
     {
         foreach (self::$achievements as $className => $instance) {
             // Check if the class actually implements Achievement interface
-            if (!($instance instanceof Achievement)) {
+            if (! ($instance instanceof Achievement)) {
                 throw new \InvalidArgumentException("Class {$className} must implement Achievement interface");
             }
 
             // Check for duplicate IDs
             $id = $instance->getId();
-            $duplicates = array_filter(self::$achievements, fn($other) => $other->getId() === $id);
+            $duplicates = array_filter(self::$achievements, fn ($other) => $other->getId() === $id);
 
             if (count($duplicates) > 1) {
                 throw new \InvalidArgumentException("Duplicate achievement ID found: {$id}");
@@ -380,7 +349,6 @@ class AchievementRegister
     /**
      * Validate that no SpecialCodeType is used by multiple SpecialAchievements.
      *
-     * @return void
      * @throws \InvalidArgumentException
      */
     private static function validateSpecialCodeTypes(): void
@@ -397,33 +365,31 @@ class AchievementRegister
                     $firstId = $specialCodeUsage[$codeValue]['achievementId'];
 
                     throw new \InvalidArgumentException(
-                        "Duplicate SpecialCodeType '{$codeValue}' found: " .
-                        "Used by both '{$firstClass}' (ID: {$firstId}) and '{$className}' (ID: {$instance->getId()}). " .
-                        "Each SpecialCodeType must be unique across all SpecialAchievements."
+                        "Duplicate SpecialCodeType '{$codeValue}' found: ".
+                        "Used by both '{$firstClass}' (ID: {$firstId}) and '{$className}' (ID: {$instance->getId()}). ".
+                        'Each SpecialCodeType must be unique across all SpecialAchievements.'
                     );
                 }
 
                 $specialCodeUsage[$codeValue] = [
                     'className' => $className,
                     'achievementId' => $instance->getId(),
-                    'instance' => $instance
+                    'instance' => $instance,
                 ];
             }
         }
 
         // Log successful validation
-        if (!empty($specialCodeUsage)) {
+        if (! empty($specialCodeUsage)) {
             Log::debug('SpecialCodeType validation passed', [
                 'special_code_types_found' => count($specialCodeUsage),
-                'codes' => array_keys($specialCodeUsage)
+                'codes' => array_keys($specialCodeUsage),
             ]);
         }
     }
 
     /**
      * Get total count of registered achievements.
-     *
-     * @return int
      */
     public static function getCount(): int
     {
